refactor(aferição): muda tipagem dos atributos
PL-3621

+++ src/main/java/br/com/zalf/prolog/webservice/v3/frota/afericao/_model/AfericaoPlacaProjection.java
@@ -5,7 +5,6 @@ import br.com.zalf.prolog.webservice.frota.pneu.afericao._model.TipoProcessoCole
 import br.com.zalf.prolog.webservice.frota.pneu.afericao.configuracao._model.FormaColetaDadosAfericaoEnum;
 import org.springframework.beans.factory.annotation.Value;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -67,20 +66,20 @@ public interface AfericaoPlacaProjection {
     Integer getPosicao();
 
     @Value("#{target.PSI}")
-    BigDecimal getPsi();
+    Double getPsi();
 
     @Value("#{target.VIDA_MOMENTO_AFERICAO}")
     Integer getVidaMomentoAfericao();
 
     @Value("#{target.ALTURA_SULCO_INTERNO}")
-    BigDecimal getAlturaSulcoInterno();
+    Double getAlturaSulcoInterno();
 
     @Value("#{target.ALTURA_SULCO_CENTRAL_INTERNO}")
-    BigDecimal getAlturaSulcoCentralInterno();
+    Double getAlturaSulcoCentralInterno();
 
     @Value("#{target.ALTURA_SULCO_CENTRAL_EXTERNO}")
-    BigDecimal getAlturaSulcoCentralExterno();
+    Double getAlturaSulcoCentralExterno();
 
     @Value("#{target.ALTURA_SULCO_EXTERNO}")
-    BigDecimal getAlturaSulcoExterno();
+    Double getAlturaSulcoExterno();
 }
